+++ source/entities/Chat/model/type.ts
@@ -6,3 +6,13 @@ export interface TChat {
   lastMessage?: string;
   usersId?: number[];
 }
+
+export interface ChatModel {
+  avatarURL?: string;
+  chatId?: number;
+  chatName?: string;
+  chatType?: "personalMessages" | "group" | "channel";
+  lastMessage?: string;
+  usersId?: number[];
+}
+export type ChatListModel = ChatModel[];
