+++ source/shared/api/types.ts
@@ -1,16 +1,6 @@
 import { TChat } from "@/entities/Chat";
 import { TUser } from "@/entities/User";
 
-export interface ChatModel {
-  avatarURL?: string;
-  chatId?: number;
-  chatName?: string;
-  chatType?: "personalMessages" | "group" | "channel";
-  lastMessage?: string;
-  usersId?: number[];
-}
-export type ChatListModel = ChatModel[];
-
 export interface ResponseError {
   error: string;
   status: string;
